# [Merged by Bors] - Speed up `Query::get_many` and add benchmarks