--- a/benches/benches/bevy_ecs/world/mod.rs
+++ b/benches/benches/bevy_ecs/world/mod.rs
@@ -27,5 +27,7 @@ criterion_group!(
     query_get_component_simple,
     query_get_component,
     query_get,
-    query_get_many2,
+    query_get_many::<2>,
+    query_get_many::<3>,
+    query_get_many::<5>,
 );