--- a/benches/benches/bevy_ecs/world/world_get.rs
+++ b/benches/benches/bevy_ecs/world/world_get.rs
@@ -406,7 +406,7 @@ pub fn query_get(criterion: &mut Criterion) {
 pub fn query_get_many<const N: usize>(criterion: &mut Criterion) {
     let mut group = criterion.benchmark_group(&format!("query_get_many_{N}"));
     group.warm_up_time(std::time::Duration::from_millis(500));
-    group.measurement_time(std::time::Duration::from_secs(4));
+    group.measurement_time(std::time::Duration::from_secs(2 * N as u64));
 
     for entity_count in RANGE.map(|i| i * 10_000) {
         group.bench_function(format!("{}_calls_table", entity_count), |bencher| {