@JoJoJet 2022-10-29
Here's another run:

| Benchmark name                       | Execution time | Change from this PR |
|--------------------------------------|----------------|---------------------|
| query_get_many_2/50000_calls_table   | 1.3157 ms      | -27.165%            |
| query_get_many_2/50000_calls_sparse  | 1.4198 ms      | -22.200%            |
| query_get_many_5/50000_calls_table   | 2.7343 ms      | -27.511%            |
| query_get_many_5/50000_calls_sparse  | 3.0702 ms      | -25.492%            |
| query_get_many_10/50000_calls_table  | 5.7574 ms      | -38.111%            |
| query_get_many_10/50000_calls_sparse | 7.9845 ms      | -26.528%            |

Criterion output: https://github.com/JoJoJet/files/blob/c4b99129978e2566b2139efb5e96462094c77aed/get-many-benches-output.zip